requirements.txt
django==5.2.7
djangorestframework==3.16.1
environs==14.3.0
djangorestframework-simplejwt==5.5.1
django-allauth==65.12.1
django-cors-headers==4.9.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
